small test

+++ otserv/branches/revmovesys/fileloader.h
@@ -78,13 +78,14 @@ protected:
 			unsigned char c = *(((unsigned char*)data) + i);
 			if(unescape && (c == NODE_START || c == NODE_END || c == ESCAPE_CHAR)) {
 				unsigned char escape = ESCAPE_CHAR;
-				int value = fwrite(&escape, 1, 1, m_file);
+				size_t value = fwrite(&escape, 1, 1, m_file);
 				if(value != 1) {
 					m_lastError = ERROR_COULDNOTWRITE;
 					return false;
 				}
 			}
-			int value = fwrite(&c, 1, 1, m_file);
+
+			size_t value = fwrite(&c, 1, 1, m_file);
 			if(value != 1) {
 				m_lastError = ERROR_COULDNOTWRITE;
 				return false;
